Ported StreamPerformanceSpec
